# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.2.0",
    "description": "Go-native Rebecca panel API. This document replaces the former FastAPI /docs page and is intended for dashboard, bot, and third-party client developers.\n\nAuthentication: protected routes use `Authorization: Bearer <token>`. The token can be the JWT returned by `POST /api/admin/token`; admin API keys are also accepted by the Go auth layer where bearer validation is used. Public subscription and home/docs routes explicitly set `security: []`.\n\nResponse compatibility: most endpoints intentionally keep the JSON shape expected by the current dashboard. Some schemas are documented as generic JSON while the migration to strict OpenAPI component schemas continues.\n\nMutation behavior: user, service, node, config, and host mutations are designed to be transactional. When runtime sync is required, Rebecca writes `node_operations` in the same transaction so failed queue writes roll back the database change."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/token": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Authenticate admin",
        "description": "Authenticate an admin and issue a JWT access token. This is the compatibility alias used by the former FastAPI OAuth2 form route.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: accepts `application/x-www-form-urlencoded` for OAuth2-compatible clients and JSON for programmatic clients.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Token"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdminToken"
      }
    },
    "/api/admin/token": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Authenticate admin",
        "description": "Authenticate an admin and issue a JWT access token.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: accepts `application/x-www-form-urlencoded` for OAuth2-compatible clients and JSON for programmatic clients.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Token"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdminToken"
      }
    },
    "/internal/admin/validate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Validate admin token",
        "description": "Internal validation endpoint for JWT and admin API key authentication.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postInternalAdminValidate"
      }
    },
    "/api/admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get current admin",
        "description": "Return the authenticated admin context, role, permissions, limits, and service assignments.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getAdmin"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create admin",
        "description": "Create a new admin if the current admin can manage admins.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdmin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or Rebecca admin API key"
      }
    },
...
    "requestBodies": {
      "JSON": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        },
        "description": "Generic JSON request body. See each operation description for required business fields and side effects."
      },
      "LoginForm": {
        "required": true,
        "content": {
          "application/x-www-form-urlencoded": {
            "schema": {
              "type": "object",
              "required": [
                "username",
                "password"
              ],
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "format": "password"
                }
              }
            }
          },
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "username",
                "password"
              ],
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "format": "password"
                }
              }
            }
          }
        },
        "description": "Admin login credentials. Form encoding is compatible with OAuth2PasswordRequestForm; JSON is accepted by the Go implementation for API clients."
      }
    },
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible Rebecca API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            }
          }
        }
      },
      "Token": {
        "description": "Admin access token.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                },
                "token_type": {
                  "type": "string",
                  "default": "bearer"
                }
              }
            }
          }
        }
      },
      "Error": {
        "description": "Error response. Most handlers return `{ \"detail\": \"...\" }`; some compatibility routes may return a plain text body when handled by net/http helpers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "additionalProperties": true
      },
      "AnyValue": {
        "nullable": true,
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      }
    }
  }
}